Explicitly expresent all the "no" instances of exceptional appositions.
The idea before was to avoid storing these, and having the default for
all words of the corresponding lemma be "no". But that would require a
table of which lemmata are supposed to be exhaustively covered by the
list of exceptions. Representing the full list is less fragile. The
exceptional appositives overlap (and should override) the
ALWAYS_PREPOSITIVE_WORDS and ALWAYS_POSTPOSITIVE_WORDS lists, so it's
important that even the "no" cases are treated as overrides.

To generate exceptional-appositives.csv:
library("tidyverse")
read_csv("Appositive ὁ, ἡ, τό.csv") |>
	select(work, book_n, line_n, word_n, word, lemma, appositive = Appositive) |>
	mutate(appositive = tolower(appositive)) |>
	filter(!(appositive %in% c("fix"))) |>
	write_csv("exceptional-appositives.csv")